isEmailAllowName: Remove redundant regex part
This regex was initially copied from another email address validation package.
It has a duplicated character set, which the first calls for "one or more" and the
second calls for "zero or more".  Considering the repeatable characters in the
first part, the second part appears to be worthless and never called.

Semver: patch